# Encoding improvements