… src/Encoding/BinaryEncoding.cs
… src/Encoding/CharEncoding.cs
… src/Encoding/DecimalEncoding.cs
… src/Encoding/DescribedEncoding.cs
… src/Encoding/DoubleEncoding.cs
… src/Encoding/FloatEncoding.cs
… src/Encoding/IntEncoding.cs
… src/Encoding/ListEncoding.cs
… src/Encoding/LongEncoding.cs
… src/Encoding/MapEncoding.cs
… src/Encoding/PrimitiveEncoding.cs
… src/Encoding/ShortEncoding.cs
… src/Encoding/SizeTracker.cs
… src/Encoding/StringEncoding.cs
… src/Encoding/SymbolEncoding.cs
… src/Encoding/TimeStampEncoding.cs
… src/Encoding/UByteEncoding.cs
… src/Encoding/UIntEncoding.cs
… src/Encoding/ULongEncoding.cs
… src/Encoding/UShortEncoding.cs
… src/Encoding/UuidEncoding.cs
… src/Framing/AmqpSequence.cs
… src/Framing/Data.cs
… src/Framing/DescribedList.cs
… src/Framing/DescribedMap.cs
… src/Framing/Frame.cs
… src/Framing/Multiple.cs
… src/Serialization/SerializableType.cs
… src/Transport/AmqpTransportInitiator.cs
… test/TestCases/AmqpCodecTests.cs