@danielmarbach 2021-12-11
Can we get those changes also somehow back ported to a minor version? That would be wonderful // cc @JoshLove-msft @jsquire